brain/pyproject.toml
[project]
name = "brain"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "langgraph>=0.3",
    "langchain-mcp-adapters>=0.1",
    "langchain-openai>=0.3",
    "python-jose[cryptography]>=3.3",
    "supabase>=2.10",
    "httpx>=0.28",
]

[tool.hatch.build.targets.wheel]
packages = ["brain"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
